@@ -1,7 +1,7 @@
-require 'factory_girl'
+require 'factory_bot'
 # Dir[Rails.root.join('spec/support/factories/*.rb')].each {|f| require f }
 Brother.delete_all
 Entry.delete_all
 
-FactoryGirl.create_list(:brother, 200)
-FactoryGirl.create_list(:entry, 100)
+FactoryBot.create_list(:brother, 200)
+FactoryBot.create_list(:entry, 100)
