@@ -1,9 +1,9 @@
-# Read about factories at https://github.com/thoughtbot/factory_girl
+# Read about factories at https://github.com/thoughtbot/factory_bot
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :vote do
-    brother_id 1
-    tshirt_id 1
-    score 1
+    brother_id { 1 }
+    tshirt_id { 1 }
+    score { 1 }
   end
 end
