@@ -6,9 +6,9 @@ describe BeansController, type: :controller do
       bro = create(:brother)
       sign_in_for_controller bro
       expect do
-        xhr :post, :create, entry_id: 1
+        post :create, xhr: true, params: { entry_id: 1 }
       end.to change(Bean, :count).by(1)
-      expect(response).to be_success
+      expect(response.response_code).to eq 200
     end
   end
 end
