@@ -5,7 +5,7 @@ describe Jinja::OsaisenController, type: :controller do
   describe "GET 'index'" do
     it "returns http success" do
       get 'index'
-      expect(response).to be_success
+      expect(response.response_code).to eq 200
     end
   end
 
