@@ -4,8 +4,8 @@ describe JinjaController, type: :controller do
 
   describe "GET 'index'" do
     it "returns http success" do
-      get 'index'
-      expect(response).to be_success
+      get :index
+      expect(response.response_code).to eq 200
     end
   end
 
